+++ tests/test_tycmd.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from unittest.mock import patch
+from subprocess import CompletedProcess
 
 import pytest
 
@@ -12,9 +13,9 @@ BLINK41_HEX = Path(__file__).parent.joinpath("blink41.hex").resolve()
 
 
 @pytest.fixture
-def mock_check_output():
-    with patch("tycmd.check_output") as mock_check_output:
-        yield mock_check_output
+def mock_run():
+    with patch("tycmd.run") as mock_run:
+        yield mock_run
 
 
 def test_identify():
@@ -26,22 +27,20 @@ def test_identify():
     assert "Teensy 4.0" in tycmd.identify(BLINK40_HEX)
 
 
-def test_list_boards(mock_check_output):
-    mock_check_output.return_value = (
+def test_list_boards(mock_run):
+    stdout = (
         '[\n  {"action": "add", "tag": "12345678-Teensy", "serial": "12345678", '
         '"description": "USB Serial", "model": "Teensy 4.1", "location": "usb-3-3", '
         '"capabilities": ["unique", "run", "rtc", "reboot", "serial"], '
         '"interfaces": [["Serial", "/dev/ttyACM0"]]}\n]\n'
     )
+    mock_run.return_value = CompletedProcess([], 0, stdout=stdout)
     output = tycmd.list_boards()
-    mock_check_output.assert_called_once_with(
-        ["tycmd", "list", "-O", "json", "-v"], text=True
-    )
     assert isinstance(output, list)
     assert isinstance(output[0], dict)
     assert output[0]["serial"] == "12345678"
 
-    mock_check_output.return_value = "[\n]\n"
+    mock_run.return_value = CompletedProcess([], 0, stdout="[\n]\n")
     output = tycmd.list_boards()
     assert isinstance(output, list)
     assert len(output) == 0
@@ -50,8 +49,8 @@ def test_list_boards(mock_check_output):
 def test_version():
     assert tycmd.version() == tycmd._TYCMD_VERSION
     with (
-        patch("tycmd.check_output", return_value="invalid") as _,
-        pytest.raises(RuntimeError),
+        patch("tycmd.run", return_value=CompletedProcess([], 0, stdout="invalid")) as _,
+        pytest.raises(ChildProcessError),
     ):
         tycmd.version()