+++ tycmd.py
@@ -1,10 +1,11 @@
 """A python wrapper for tycmd."""
 
 from pathlib import Path
-from subprocess import check_output, CalledProcessError, STDOUT
+from subprocess import CalledProcessError, PIPE, run
 import json
 import re
 from logging import getLogger
+from typing import Literal, IO
 
 __version__ = "0.2.0"
 _TYCMD_VERSION = "0.9.9"
@@ -12,7 +13,14 @@ _TYCMD_VERSION = "0.9.9"
 log = getLogger(__name__)
 
 
-def upload(filename: Path | str, port: str | None = None, serial: str | None = None):
+def upload(
+    filename: Path | str,
+    port: str | None = None,
+    serial: str | None = None,
+    check: bool = True,
+    reset_board: bool = True,
+    rtc_mode: Literal["local", "utc", "none"] = "local",
+):
     """
     Upload firmware to board.
 
@@ -26,14 +34,24 @@ def upload(filename: Path | str, port: str | None = None, serial: str | None = N
 
     serial : str, optional
         Serial number of targeted board.
+
+    check : bool, optional
+        Check if board is compatible before upload. Defaults to True.
+
+    reset_board : bool, optional
+        Reset the device once the upload is finished. Defaults to True.
+
+    rtc_mode : str, optional
+        Set RTC if supported: 'local' (default), 'utc' or 'none'.
     """
     filename = str(_parse_firmware_file(filename))
-    try:
-        _call_tycmd(["upload", filename], port=port, serial=serial)
-    except CalledProcessError:
-        # TODO: implementation ...
-        # print(e.stdout)
-        pass
+    args = ["upload"]
+    if not check:
+        args.append("--nocheck")
+    if not reset_board:
+        args.append("--noreset")
+    args.extend(["--rtc", rtc_mode, filename])
+    _call_tycmd(args, port=port, serial=serial, stdout=None)
 
 
 def identify(filename: Path | str) -> list[str]:
@@ -51,7 +69,7 @@ def identify(filename: Path | str) -> list[str]:
         List of models compatible with firmware.
     """
     filename = str(_parse_firmware_file(filename))
-    return_string = _call_tycmd(args=["identify", filename, "--json", "-qqq"])
+    return_string = _call_tycmd(args=["identify", filename, "--json"])
     return_string = return_string.replace("\\", "\\\\")
     output = json.loads(return_string)
     if "error" in output:
@@ -68,8 +86,9 @@ def list_boards() -> list[dict]:
     list[dict]
         List of available devices.
     """
-    args = ["tycmd", "list", "-O", "json", "-v"]
-    return_string = check_output(args, text=True)
+    args = ["list", "-O", "json", "-v"]
+    return_string = _call_tycmd(args)
+    # return_string = check_output(args, text=True)
     return json.loads(return_string)
 
 
@@ -90,7 +109,7 @@ def version() -> str:
     output = _call_tycmd(["--version"])
     match = re.search(r"\d+\.\d+\.\d+", output)
     if match is None:
-        raise RuntimeError("Could not determine tycmd version")
+        raise ChildProcessError("Could not determine tycmd version")
     else:
         return match.group()
 
@@ -142,14 +161,26 @@ def _call_tycmd(
     serial: str | None = None,
     family: str | None = None,
     port: str | None = None,
+    stdout: None | int | IO = PIPE,
+    raise_on_stderr: bool = False,
 ) -> str:
     tag = _assemble_tag(serial=serial, family=family, port=port)
-    args = ["tycmd"] + args + tag
-    # log.debug(f"Starting subprocess: {' '.join(args)}")
-    # with Popen(args, stdout=PIPE, stderr=PIPE, bufsize=1, text=True) as proc:
-    #     for line in proc.stdout:
-    #         print(line, end='')
-    return check_output(args, text=True, stderr=STDOUT)
+    args = ["tycmd"] + tag + args
+    log.debug(f"Calling subprocess: {' '.join(args)}")
+
+    # call tycmd and raise non-zero exit codes as a RuntimeError
+    try:
+        p = run(args, stdout=stdout, stderr=PIPE, check=True, text=True)
+    except CalledProcessError as e:
+        raise ChildProcessError(str.strip(e.stderr or "")) from e
+
+    # tycmd doesn't always set a non-negative exit code when an error occurs. If
+    # raise_on_stderr is True and the subprocess' stderr is not None we'll still raise a
+    # ChildProcessError despite the exit code being 0.
+    if raise_on_stderr and isinstance(p.stderr, str):
+        raise ChildProcessError(p.stderr.strip())
+
+    return p.stdout.strip() if isinstance(p.stdout, str) else ""
 
 
 def _assemble_tag(